requirements.txt
einops
numpy
opencv-python
Pillow
PyYAML
scikit-learn
scipy
tabulate
tensorboardX
torch
torchvision
easydict